requirements.txt
numpy
pillow
opencv
flask-swagger-ui
flask
enum

